requirements.txt
flask
ccxt
python-dotenv
